requirements.txt
requests>=2.25.1python-dotenv>=1.0.0
python-dotenv>=1.0.0
